+++ packages/dso-toolkit/src/components/card/card.models.ts
@@ -1,4 +1,4 @@
-import { Anchor } from "../anchor/anchor.models.js";
+import { Link } from "../link/link.models.js";
 import { Button } from "../button/button.models.js";
 import { Label } from "../label/label.models.js";
 import { Selectable } from "../selectable/selectable.models.js";
@@ -9,7 +9,7 @@ export interface Card<TemplateFnReturnType> {
   label: string;
   href: string;
   active?: boolean;
-  mode?: Anchor["mode"];
+  mode?: Link["mode"];
   selectable?: Selectable<TemplateFnReturnType>;
   content?: TemplateFnReturnType;
   interactions?: Array<Button | Label | Toggletip<TemplateFnReturnType> | SlideToggle>;