+++ storybook/src/components/modal/modal.content.ts
@@ -51,7 +51,7 @@ export function passiveFooter({ buttonTemplate }: Templates) {
   return buttonTemplate({ label: "Sluiten", variant: "tertiary", type: "button" });
 }
 
-export function confirmBody({ anchorTemplate }: Templates) {
+export function confirmBody({ linkTemplate }: Templates) {
   return html`
     <p>
       Welkom op het nieuwe Omgevingsloket. Deze website is nog in ontwikkeling. Dat betekent dat bepaalde onderdelen nog
@@ -59,7 +59,7 @@ export function confirmBody({ anchorTemplate }: Templates) {
       kaart. Ook kunt u nog geen aanvraag of melding indienen via dit loket.
     </p>
     <p>
-      De website ${anchorTemplate({ label: "Aan de slag met de omgevingswet", url: "#" })} biedt extra uitleg over het
+      De website ${linkTemplate({ label: "Aan de slag met de omgevingswet", url: "#" })} biedt extra uitleg over het
       Omgevingsloket en de ontwikkeling ervan. Ook vindt u informatie over hoe u kunt oefenen met aansluiten van uw
       lokale systemen op het loket.
     </p>