+++ storybook/src/components/viewer-grid/viewer-grid.core-stories.ts
@@ -80,7 +80,7 @@ const {
   templateContainer,
   storyTemplates: (templates) => {
     const {
-      anchorTemplate,
+      linkTemplate,
       badgeTemplate,
       viewerGridTemplate,
       alertTemplate,
@@ -128,7 +128,7 @@ const {
                 term: html`Besluit:`,
                 descriptions: [
                   {
-                    content: anchorTemplate({
+                    content: linkTemplate({
                       label: "Bekijk besluit",
                       url: "#",
                       icon: {
@@ -245,7 +245,7 @@ const {
           children: html`<p>
               Ut elit purus, scelerisque nec tincidunt id, dictum at sapien. Nulla at felis quam. Nullam commodo ex
               ultrices, viverra urna a, pretium arcu. Nunc eget cursus lorem. Sed massa nunc, maximus sodales
-              ${anchorTemplate({ label: "ultrices", url: "#" })} nec, luctus et lectus. Interdum et malesuada fames ac
+              ${linkTemplate({ label: "ultrices", url: "#" })} nec, luctus et lectus. Interdum et malesuada fames ac
               ante ipsum primis in faucibus. Morbi ultrices tincidunt ipsum, sit amet ultricies nulla pulvinar nec. Cras
               sed tellus in nunc viverra aliquam. Aenean sed libero nulla. Curabitur placerat ullamcorper nisl, ut
               facilisis tortor rhoncus a. Etiam vel ex nec eros porttitor aliquam. Duis blandit vel ex at venenatis.
@@ -257,11 +257,11 @@ const {
               Ut elit purus, scelerisque nec tincidunt id, dictum at sapien. Nulla at felis quam. Nullam commodo ex
               ultrices, viverra urna a, pretium arcu. Nunc eget cursus lorem. Sed massa nunc, maximus sodales ultrices
               nec, luctus et lectus. Interdum et malesuada fames ac ante
-              ${anchorTemplate({ label: "ipsum primis", url: "#" })} in faucibus. Morbi ultrices tincidunt ipsum, sit
-              amet ultricies nulla pulvinar nec. Cras sed tellus in nunc viverra aliquam. Aenean sed libero nulla.
-              Curabitur placerat ullamcorper nisl, ut facilisis tortor rhoncus a. Etiam vel ex nec eros porttitor
-              aliquam. Duis blandit vel ex at venenatis. Pellentesque habitant morbi tristique senectus et netus et
-              malesuada fames ac turpis egestas. Nulla sodales facilisis hendrerit.
+              ${linkTemplate({ label: "ipsum primis", url: "#" })} in faucibus. Morbi ultrices tincidunt ipsum, sit amet
+              ultricies nulla pulvinar nec. Cras sed tellus in nunc viverra aliquam. Aenean sed libero nulla. Curabitur
+              placerat ullamcorper nisl, ut facilisis tortor rhoncus a. Etiam vel ex nec eros porttitor aliquam. Duis
+              blandit vel ex at venenatis. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac
+              turpis egestas. Nulla sodales facilisis hendrerit.
             </p>`,
         })}`,
       },