+++ storybook/src/example-pages/toepassingen/stelselcatalogus/begrippen.content.ts
@@ -34,7 +34,7 @@ export const mainMenu = [
   },
 ];
 
-export function definitions1({ anchorTemplate, toggletipTemplate }: Templates): Definition<TemplateResult>[] {
+export function definitions1({ linkTemplate, toggletipTemplate }: Templates): Definition<TemplateResult>[] {
   return [
     {
       term: html`
@@ -74,14 +74,14 @@ export function definitions1({ anchorTemplate, toggletipTemplate }: Templates): 
       term: html` Valt binnen<span class="sr-only">:</span> `,
       descriptions: [
         {
-          content: anchorTemplate({ label: "Aquo begrippenkader", url: "#", mode: "extern" }),
+          content: linkTemplate({ label: "Aquo begrippenkader", url: "#", mode: "extern" }),
         },
       ],
     },
   ];
 }
 
-export function definitions2({ toggletipTemplate, anchorTemplate }: Templates): Definition<TemplateResult>[] {
+export function definitions2({ toggletipTemplate, linkTemplate }: Templates): Definition<TemplateResult>[] {
   return [
     {
       term: html`
@@ -90,7 +90,7 @@ export function definitions2({ toggletipTemplate, anchorTemplate }: Templates): 
       `,
       descriptions: [
         {
-          content: anchorTemplate({
+          content: linkTemplate({
             label: "https://water.omgevingswet.overheid.nl/aquobegrippenkader/id/concept/Aanlegplaats",
             url: "#",
             mode: "extern",
@@ -101,7 +101,7 @@ export function definitions2({ toggletipTemplate, anchorTemplate }: Templates): 
   ];
 }
 
-export function definitions3({ toggletipTemplate, anchorTemplate }: Templates): Definition<TemplateResult>[] {
+export function definitions3({ toggletipTemplate, linkTemplate }: Templates): Definition<TemplateResult>[] {
   return [
     {
       term: html`
@@ -110,14 +110,14 @@ export function definitions3({ toggletipTemplate, anchorTemplate }: Templates): 
       `,
       descriptions: [
         {
-          content: anchorTemplate({ label: "Aanlegsteiger", url: "#", mode: "extern" }),
+          content: linkTemplate({ label: "Aanlegsteiger", url: "#", mode: "extern" }),
         },
       ],
     },
   ];
 }
 
-export function definitions4({ toggletipTemplate, anchorTemplate }: Templates): Definition<TemplateResult>[] {
+export function definitions4({ toggletipTemplate, linkTemplate }: Templates): Definition<TemplateResult>[] {
   return [
     {
       term: html`
@@ -126,14 +126,14 @@ export function definitions4({ toggletipTemplate, anchorTemplate }: Templates): 
       `,
       descriptions: [
         {
-          content: anchorTemplate({ label: "Steiger plaatsen", url: "#", mode: "extern" }),
+          content: linkTemplate({ label: "Steiger plaatsen", url: "#", mode: "extern" }),
         },
       ],
     },
   ];
 }
 
-export function definitions5({ toggletipTemplate, anchorTemplate }: Templates): Definition<TemplateResult>[] {
+export function definitions5({ toggletipTemplate, linkTemplate }: Templates): Definition<TemplateResult>[] {
   return [
     {
       term: html`
@@ -142,7 +142,7 @@ export function definitions5({ toggletipTemplate, anchorTemplate }: Templates): 
       `,
       descriptions: [
         {
-          content: anchorTemplate({ label: "Aquo", url: "#", mode: "extern" }),
+          content: linkTemplate({ label: "Aquo", url: "#", mode: "extern" }),
         },
       ],
     },
@@ -153,7 +153,7 @@ export function definitions5({ toggletipTemplate, anchorTemplate }: Templates): 
       `,
       descriptions: [
         {
-          content: anchorTemplate({
+          content: linkTemplate({
             label: "https://www.w3.org/2004/02/skos/core#Concept",
             url: "#",
             mode: "extern",
@@ -168,7 +168,7 @@ export function definitions5({ toggletipTemplate, anchorTemplate }: Templates): 
       `,
       descriptions: [
         {
-          content: anchorTemplate({
+          content: linkTemplate({
             label: "https://water.omgevingswet.overheid.nl/aquabegrippenkader/doc/20180412000000/concept/Aanlegplaats",
             url: "#",
             mode: "extern",